Move getVariantPresets to AssetVariantGenerator
