+++ Neos.Media/Tests/Unit/Domain/Service/AssetServiceTest.php
@@ -60,32 +60,4 @@ class AssetServiceTest extends UnitTestCase
         $repository = $mockAssetService->_call('getRepository', $mockAsset);
         $this->assertInstanceOf($expectedRepositoryClassName, $repository);
     }
-
-    /**
-     * @test
-     */
-    public function getVariantPresetsReturnsConfiguration(): void
-    {
-        $configuration = [
-            'Flownative.Demo:Preset1' => [
-                'label' => 'Demo Preset 1',
-                'mediaTypePatterns' => ['~image/.*~'],
-                'variants' => []
-            ],
-            'Flownative.Demo:Preset2' => [
-                'label' => 'Demo Preset 2',
-                'mediaTypePatterns' => ['~image/.*~'],
-                'variants' => []
-            ]
-        ];
-
-        $assetService = new AssetService();
-        $this->inject($assetService, 'variantPresetsConfiguration', $configuration);
-
-        $presets = $assetService->getVariantPresets();
-        self::assertArrayHasKey('Flownative.Demo:Preset1', $presets);
-        self::assertSame($configuration['Flownative.Demo:Preset1']['label'], (string)$presets['Flownative.Demo:Preset1']->label());
-        self::assertArrayHasKey('Flownative.Demo:Preset2', $presets);
-        self::assertSame($configuration['Flownative.Demo:Preset2']['label'], (string)$presets['Flownative.Demo:Preset2']->label());
-    }
 }
